requirements.txt
# requirements.txt WPGeminiPublisher
google-generativeai   
requests              
python-markdown       
Markdown              
customtkinter         
